+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/Components.java
@@ -18,9 +18,7 @@
  */
 package is.codion.swing.common.ui.component;
 
-import is.codion.common.item.Item;
 import is.codion.common.value.ValueSet;
-import is.codion.swing.common.model.component.combobox.FilterComboBoxModel;
 import is.codion.swing.common.ui.component.builder.ComponentBuilder;
 import is.codion.swing.common.ui.component.button.ButtonBuilder;
 import is.codion.swing.common.ui.component.button.ButtonPanelBuilder;
@@ -74,7 +72,6 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.time.OffsetDateTime;
 import java.time.temporal.Temporal;
-import java.util.List;
 
 import static is.codion.swing.common.model.component.combobox.FilterComboBoxModel.booleanItems;
 import static is.codion.swing.common.ui.layout.Layouts.*;
@@ -142,36 +139,17 @@ public final class Components {
 	}
 
 	/**
-	 * @return a boolean based JComboBox builder
+	 * @return a boolean based {@link ItemComboBoxBuilder}
 	 */
 	public static ItemComboBoxBuilder<Boolean> booleanComboBox() {
-		return ItemComboBoxBuilder.builder(booleanItems());
+		return ItemComboBoxBuilder.builder().items(booleanItems());
 	}
 
 	/**
-	 * @param comboBoxModel the combo box model
-	 * @return a boolean based JComboBox builder
+	 * @return a {@link ItemComboBoxBuilder.BuilderFactory}
 	 */
-	public static ItemComboBoxBuilder<Boolean> booleanComboBox(FilterComboBoxModel<Item<Boolean>> comboBoxModel) {
-		return ItemComboBoxBuilder.builder(comboBoxModel);
-	}
-
-	/**
-	 * @param comboBoxModel the combo box model
-	 * @param <T> the value type
-	 * @return a {@link Item} based JComboBox builder
-	 */
-	public static <T> ItemComboBoxBuilder<T> itemComboBox(FilterComboBoxModel<Item<T>> comboBoxModel) {
-		return ItemComboBoxBuilder.builder(comboBoxModel);
-	}
-
-	/**
-	 * @param values the values
-	 * @param <T> the value type
-	 * @return a {@link Item} based JComboBox builder
-	 */
-	public static <T> ItemComboBoxBuilder<T> itemComboBox(List<Item<T>> values) {
-		return ItemComboBoxBuilder.builder(values);
+	public static ItemComboBoxBuilder.BuilderFactory itemComboBox() {
+		return ItemComboBoxBuilder.builder();
 	}
 
 	/**